requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
thefuzz>=0.22.0
rapidfuzz>=3.0.0
requests>=2.31.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
google-generativeai>=0.3.0
aiohttp>=3.9.0
streamlit-autorefresh>=1.0.1
